Add notice about the state of the project.
The BinNavi project is no longer under active development.

As the team's core responsibilies shifted, our internal use cases
basically became non-existent.

We thus failed to deliver adequate 64-bit support and necessary
improvements, despite some encouraging initial participation from the
community.

So, in the interest of being transparent and honest, I added a note
about the current state of the project to the README.

Anyone in the infosec and RE community is of cource welcome to fork this
project and to continue working on it. A good starting point to get this
up to speed again, is to look at the `reil-x64` branch.

I've also closed out all issues and PRs, but nothing has been deleted
and should be available for future reference.

Thanks everyone who sent us kind words, PRs and/or bug reports.

Signed-off-by: Christian Blichmann <mail@blichmann.eu>